// src/app/shared/enums/enums.ts
export enum AnnotationType {
    Text = "Text",
    Image = "Image",
}

export enum AnnotationActions {
    resize = "resize",
    move = "move",
    focus = "focus",
}
export enum ViewActions {
    plusScale = "plusScale",
    minusScale = "minusScale",
    addText = "addText",
    addImage = "addImage",
    save = "save",
    previousStep = "previousStep",
    nextStep = "nextStep",
}